requirements.txt
flask==2.3.2
flask-cors==3.0.10
dnspython==2.3.0
aiosmtpd==1.4.4.post2
gunicorn==20.1.0
aiohttp==3.9.5
pyOpenSSL>=23.2.0
python-dotenv==1.0.0
pytest==7.4.0
